docs(diffusion): tighten the kitchen_int8 section and fix its backend claims
The Kitchen INT8 section had grown into a benchmark report: a launch-bound
analysis, kernel geomeans and a paragraph about a later change no reader of
this page can see. It now states what a user needs (what runs where, what
stays BF16, memory at load) and one measurement table.

Two claims were wrong against the code: `auto` was described as using
comfy_kitchen "where that package is installed", but loading stops with an
install hint when it is missing (only the output-width fallback checks the
package), and the MiniMax-H3 cookbook said the sgl-kernel ops cover
"CC 9.0 and later" although CC 10.3 is refused by design. The serialized
ConvRot INT8 row now lists sgl-kernel as a dependency option too, matching
its notes.

+++ python/sglang/multimodal_gen/test/unit/test_kitchen_int8_backends.py
@@ -348,7 +348,7 @@ def test_fp16_parameters_are_cast_for_the_kernel_in_every_helper():
     x_bf16 = x.to(torch.bfloat16)
     apply_convrot_int8_shared_input_out(x=x_bf16, layers=layers, outs=outs)
     for out, layer in zip(outs, layers, strict=True):
-        assert torch.equal(out, layer(x_bf16)[0].to(torch.bfloat16))
+        assert torch.equal(out, layer(x_bf16)[0])
 
 
 @requires_kernel
